c2

+++ src/Main.java
@@ -1,10 +1,9 @@
 public class Main {
 
-
     public static void main(String[] args) {
         Calculator calc = new Calculator();
-        int x = 2;
-        int y = 5;
+        float x = 2;
+        float y = 5;
 
         float a = calc.sum(x, y);
         System.out.println(a);
